# backend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LumiCreate-Pro API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/audio-engine/generate-stream": {
      "post": {
        "tags": [
          "audio-engine"
        ],
        "summary": "Generate Single Stream",
        "operationId": "generate_single_stream_api_audio_engine_generate_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/routers__audio_engine__SingleGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audio-engine/generate-batch-stream": {
      "post": {
        "tags": [
          "audio-engine"
        ],
        "summary": "Generate Batch Stream",
        "operationId": "generate_batch_stream_api_audio_engine_generate_batch_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/routers__audio_engine__BatchGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audio-engine/stitch-scene": {
      "post": {
        "tags": [
          "audio-engine"
        ],
        "summary": "Stitch Scene",
        "description": "Merge clips for one scene into a single WAV (with pre/post silence).",
        "operationId": "stitch_scene_api_audio_engine_stitch_scene_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audio-engine/ms-tts": {
      "post": {
        "tags": [
          "audio-engine"
        ],
        "summary": "Ms Tts",
        "description": "Generate a single TTS clip with Microsoft Edge neural voices.\nReturns base64 MP3 by default; pass format=\"wav\" for stitch-scene compatible WAV.",
        "operationId": "ms_tts_api_audio_engine_ms_tts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MsTtsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audio-engine/ms-tts/test-all": {
      "post": {
        "tags": [
          "audio-engine"
        ],
        "summary": "Ms Tts Test All",
        "description": "逐一测试 Microsoft Edge neural voices；返回 {voice: ok/err}，\n并将通过的 voice 列表写入 settings.audio_engine.msedge_available_voices。",
        "operationId": "ms_tts_test_all_api_audio_engine_ms_tts_test_all_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestAllRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchDialogue": {
        "properties": {
          "scene_id": {
            "type": "string",
            "title": "Scene Id"
          },
          "dialogue_id": {
            "type": "string",
            "title": "Dialogue Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "voice_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Ref"
          },
          "emo_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emo Ref"
          },
          "emo_weight": {
            "type": "number",
            "title": "Emo Weight",
            "default": 0.8
          },
          "lang": {
            "type": "string",
            "title": "Lang",
            "default": "zh"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "msedge_voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msedge Voice"
          },
          "msedge_rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msedge Rate"
          }
        },
        "type": "object",
        "required": [
          "scene_id",
          "dialogue_id",
          "text"
        ],
        "title": "BatchDialogue"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MsTtsRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "voice": {
            "type": "string",
            "title": "Voice",
            "default": "zh-CN-XiaoxiaoNeural"
          },
          "rate": {
            "type": "string",
            "title": "Rate",
            "default": "+0%"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "mp3"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "MsTtsRequest"
      },
...
      "StitchClip": {
        "properties": {
          "data": {
            "type": "string",
            "title": "Data"
          },
          "pre_silence_ms": {
            "type": "integer",
            "title": "Pre Silence Ms",
            "default": 0
          },
          "post_silence_ms": {
            "type": "integer",
            "title": "Post Silence Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StitchClip"
      },
      "StitchRequest": {
        "properties": {
          "clips": {
            "items": {
              "$ref": "#/components/schemas/StitchClip"
            },
            "type": "array",
            "title": "Clips"
          }
        },
        "type": "object",
        "required": [
          "clips"
        ],
        "title": "StitchRequest"
      },
...
      "TestAllRequest": {
        "properties": {
          "voices": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voices"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": "测试"
          },
          "save": {
            "type": "boolean",
            "title": "Save",
            "default": true
          }
        },
        "type": "object",
        "title": "TestAllRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "routers__audio_engine__BatchGenerateRequest": {
        "properties": {
          "gen_count": {
            "type": "integer",
            "title": "Gen Count",
            "default": 3
          },
          "speed": {
            "type": "number",
            "title": "Speed",
            "default": 1.0
          },
          "dialogues": {
            "items": {
              "$ref": "#/components/schemas/BatchDialogue"
            },
            "type": "array",
            "title": "Dialogues"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "dialogues"
        ],
        "title": "BatchGenerateRequest"
      },
...
      "routers__audio_engine__SingleGenerateRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "voice_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Ref"
          },
          "emo_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emo Ref"
          },
          "emo_weight": {
            "type": "number",
            "title": "Emo Weight",
            "default": 0.8
          },
          "lang": {
            "type": "string",
            "title": "Lang",
            "default": "zh"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "speed": {
            "type": "number",
            "title": "Speed",
            "default": 1.0
          },
          "scene_id": {
            "type": "string",
            "title": "Scene Id",
            "default": ""
          },
          "dialogue_id": {
            "type": "string",
            "title": "Dialogue Id",
            "default": ""
          },
          "slot_index": {
            "type": "integer",
            "title": "Slot Index",
            "default": 0
          },
          "msedge_voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msedge Voice"
          },
          "msedge_rate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msedge Rate"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SingleGenerateRequest"
      },
...
    }
  }
}